Move RSA key module to its own package
