+++ pkg/keys/keys_test/rsa_test.go
@@ -1,20 +1,20 @@
-package jwt_test
+package keys_test
 
 import (
-	"eau-de-go/pkg/jwt_util"
+	"eau-de-go/pkg/keys"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
 func TestGetInMemoryRsaKeyPair_Singleton(t *testing.T) {
-	keyPair1 := jwt_util.GetInMemoryRsaKeyPair()
-	keyPair2 := jwt_util.GetInMemoryRsaKeyPair()
+	keyPair1 := keys.GetInMemoryRsaKeyPair()
+	keyPair2 := keys.GetInMemoryRsaKeyPair()
 
 	assert.Equal(t, keyPair1, keyPair2, "GetInMemoryRsaKeyPair should always return the same instance")
 }
 
 func TestInMemoryRsaKeyPair_GetVerificationKey(t *testing.T) {
-	keyPair := jwt_util.GetInMemoryRsaKeyPair()
+	keyPair := keys.GetInMemoryRsaKeyPair()
 
 	verificationKey1, err1 := keyPair.GetVerificationKey()
 	assert.Nil(t, err1, "GetVerificationKey should not return an error")
@@ -26,7 +26,7 @@ func TestInMemoryRsaKeyPair_GetVerificationKey(t *testing.T) {
 }
 
 func TestInMemoryRsaKeyPair_GetSigningKey(t *testing.T) {
-	keyPair := jwt_util.GetInMemoryRsaKeyPair()
+	keyPair := keys.GetInMemoryRsaKeyPair()
 
 	signingKey1, err1 := keyPair.GetSigningKey()
 	assert.Nil(t, err1, "GetSigningKey should not return an error")
